requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
idna==3.6
numpy==1.26.4
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sendgrid==6.11.0
sqlparse==0.4.4
twilio==8.12.0
typing_extensions==4.9.0
urllib3==2.2.0
whitenoise==6.6.0
